# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "TaxVault API",
                 "version":  "1.1.0"
             },
    "paths":  {
...
                  "/api/v1/tax-profiles":  {
                                               "post":  {
                                                            "tags":  [
                                                                         "Tax Profiles",
                                                                         "Tax Profiles"
                                                                     ],
                                                            "summary":  "Create",
                                                            "operationId":  "create_api_v1_tax_profiles_post",
                                                            "requestBody":  {
                                                                                "content":  {
                                                                                                "application/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/TaxProfileCreate"
                                                                                                                                    }
                                                                                                                     }
                                                                                            },
                                                                                "required":  true
                                                                            },
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Successful Response",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {

                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      },
                                                                              "422":  {
                                                                                          "description":  "Validation Error",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                           },
                  "/api/v1/tax-profiles/{client_id}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Tax Profiles",
                                                                                    "Tax Profiles"
                                                                                ],
                                                                       "summary":  "Get",
                                                                       "operationId":  "get_api_v1_tax_profiles__client_id__get",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "client_id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "title":  "Client Id"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Successful Response",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {

                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 },
                                                                                         "422":  {
                                                                                                     "description":  "Validation Error",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
                  "/api/v1/tax-profiles/{id}":  {
                                                    "put":  {
                                                                "tags":  [
                                                                             "Tax Profiles",
                                                                             "Tax Profiles"
                                                                         ],
                                                                "summary":  "Update",
                                                                "operationId":  "update_api_v1_tax_profiles__id__put",
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "id",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "integer",
                                                                                                      "title":  "Id"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "requestBody":  {
                                                                                    "required":  true,
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/TaxProfileUpdate"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "Successful Response",
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {

                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          }
                                                                                          },
                                                                                  "422":  {
                                                                                              "description":  "Validation Error",
                                                                                              "content":  {
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            },
                                                    "delete":  {
                                                                   "tags":  [
                                                                                "Tax Profiles",
                                                                                "Tax Profiles"
                                                                            ],
                                                                   "summary":  "Delete",
                                                                   "operationId":  "delete_api_v1_tax_profiles__id__delete",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "title":  "Id"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {

                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                }
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "TaxProfileCreate":  {
                                                                "properties":  {
                                                                                   "salary_income":  {
                                                                                                         "type":  "boolean",
                                                                                                         "title":  "Salary Income",
                                                                                                         "default":  false
                                                                                                     },
                                                                                   "business_income":  {
                                                                                                           "type":  "boolean",
                                                                                                           "title":  "Business Income",
                                                                                                           "default":  false
                                                                                                       },
                                                                                   "house_property":  {
                                                                                                          "type":  "boolean",
                                                                                                          "title":  "House Property",
                                                                                                          "default":  false
                                                                                                      },
                                                                                   "capital_gain":  {
                                                                                                        "type":  "boolean",
                                                                                                        "title":  "Capital Gain",
                                                                                                        "default":  false
                                                                                                    },
                                                                                   "other_sources":  {
                                                                                                         "type":  "boolean",
                                                                                                         "title":  "Other Sources",
                                                                                                         "default":  false
                                                                                                     },
                                                                                   "foreign_assets":  {
                                                                                                          "type":  "boolean",
                                                                                                          "title":  "Foreign Assets",
                                                                                                          "default":  false
                                                                                                      },
                                                                                   "regime":  {
                                                                                                  "type":  "string",
                                                                                                  "title":  "Regime",
                                                                                                  "default":  "NEW"
                                                                                              },
                                                                                   "financial_year":  {
                                                                                                          "type":  "string",
                                                                                                          "title":  "Financial Year",
                                                                                                          "default":  "2026-27"
                                                                                                      },
                                                                                   "client_id":  {
                                                                                                     "type":  "integer",
                                                                                                     "title":  "Client Id"
                                                                                                 }
                                                                               },
                                                                "type":  "object",
                                                                "required":  [
                                                                                 "client_id"
                                                                             ],
                                                                "title":  "TaxProfileCreate"
                                                            },
                                       "TaxProfileUpdate":  {
                                                                "properties":  {
                                                                                   "salary_income":  {
                                                                                                         "type":  "boolean",
                                                                                                         "title":  "Salary Income",
                                                                                                         "default":  false
                                                                                                     },
                                                                                   "business_income":  {
                                                                                                           "type":  "boolean",
                                                                                                           "title":  "Business Income",
                                                                                                           "default":  false
                                                                                                       },
                                                                                   "house_property":  {
                                                                                                          "type":  "boolean",
                                                                                                          "title":  "House Property",
                                                                                                          "default":  false
                                                                                                      },
                                                                                   "capital_gain":  {
                                                                                                        "type":  "boolean",
                                                                                                        "title":  "Capital Gain",
                                                                                                        "default":  false
                                                                                                    },
                                                                                   "other_sources":  {
                                                                                                         "type":  "boolean",
                                                                                                         "title":  "Other Sources",
                                                                                                         "default":  false
                                                                                                     },
                                                                                   "foreign_assets":  {
                                                                                                          "type":  "boolean",
                                                                                                          "title":  "Foreign Assets",
                                                                                                          "default":  false
                                                                                                      },
                                                                                   "regime":  {
                                                                                                  "type":  "string",
                                                                                                  "title":  "Regime",
                                                                                                  "default":  "NEW"
                                                                                              },
                                                                                   "financial_year":  {
                                                                                                          "type":  "string",
                                                                                                          "title":  "Financial Year",
                                                                                                          "default":  "2026-27"
                                                                                                      }
                                                                               },
                                                                "type":  "object",
                                                                "title":  "TaxProfileUpdate"
                                                            },
...
                                   },
                       "securitySchemes":  {
                                               "OAuth2PasswordBearer":  {
                                                                            "type":  "oauth2",
                                                                            "flows":  {
                                                                                          "password":  {
                                                                                                           "scopes":  {

                                                                                                                      },
                                                                                                           "tokenUrl":  "/api/v1/auth/login"
                                                                                                       }
                                                                                      }
                                                                        }
                                           }
                   }
}